+++ ext/monitor/src/main/java/app/AlertModule.java
@@ -1,15 +1,14 @@
 package app;
 
 import app.monitor.AlertConfig;
-import app.monitor.PagerdutyConfig;
 import app.monitor.alert.AlertService;
 import app.monitor.channel.Channel;
 import app.monitor.channel.ChannelManager;
+import app.monitor.channel.PagerDutyClient;
 import app.monitor.channel.SlackClient;
 import app.monitor.kafka.ActionLogMessageHandler;
 import app.monitor.kafka.EventMessageHandler;
 import app.monitor.kafka.StatMessageHandler;
-import app.monitor.channel.PagerdutyClient;
 import core.framework.http.HTTPClient;
 import core.framework.json.Bean;
 import core.framework.log.message.ActionLogMessage;
@@ -36,33 +35,24 @@ public class AlertModule extends Module {
 
     private void configureChannels() {
         Map<String, Channel> channels = new HashMap<>();
-        configureSlackChannel(channels);
-        configurePagerdutyChannel(channels);
+        HTTPClient httpClient = HTTPClient.builder()
+            .maxRetries(3)
+            .retryWaitTime(Duration.ofSeconds(2))    // slack has rate limit with 1 message per second, here to slow down further when hit limit, refer to https://api.slack.com/docs/rate-limits
+            .build();
+        configureSlackChannel(channels, httpClient);
+        configurePagerDutyChannel(channels, httpClient);
         bind(new ChannelManager(channels, "slack"));
     }
 
-    private void configureSlackChannel(Map<String, Channel> channels) {
-        property("app.slack.token").ifPresent((String slackToken) -> {
-            HTTPClient httpClient = HTTPClient.builder()
-                .maxRetries(3)
-                .retryWaitTime(Duration.ofSeconds(2))   // slack has rate limit with 1 message per second, here to slow down further when hit limit, refer to https://api.slack.com/docs/rate-limits
-                .build();
-            channels.put("slack", new SlackClient(httpClient, slackToken));
-        });
+    private void configureSlackChannel(Map<String, Channel> channels, HTTPClient httpClient) {
+        property("app.slack.token").ifPresent((String token) ->
+            channels.put("slack", new SlackClient(httpClient, token)));
     }
 
-    private void configurePagerdutyChannel(Map<String, Channel> channels) {
-        Optional<String> fromOpt = property("app.pagerduty.from");
-        property("app.pagerduty.token").ifPresent((String token) -> {
-            PagerdutyConfig pdConfig = new PagerdutyConfig();
-            pdConfig.token = token;
-            pdConfig.from = fromOpt.orElseThrow(() -> new IllegalArgumentException("app.pagerduty.from must not be null when app.pagerduty.token is not null"));
-            HTTPClient httpClient = HTTPClient.builder()
-                .maxRetries(3)
-                .retryWaitTime(Duration.ofSeconds(30))
-                .build();
-            channels.put("pagerduty", new PagerdutyClient(httpClient, pdConfig));
-        });
+    private void configurePagerDutyChannel(Map<String, Channel> channels, HTTPClient httpClient) {
+        Optional<String> from = property("app.pagerduty.from");
+        property("app.pagerduty.token").ifPresent((String token) ->
+            channels.put("pagerduty", new PagerDutyClient(httpClient, token, from.orElseThrow())));
     }
 
     private void configureAlert(String alertConfig) {