+++ ext/monitor/src/main/java/app/monitor/PagerdutyConfig.java
@@ -1,9 +1,0 @@
-package app.monitor;
-
-/**
- * @author ajax
- */
-public class PagerdutyConfig {
-    public String token;
-    public String from;
-}