+++ ext/monitor/src/main/java/app/monitor/channel/ChannelManager.java
@@ -19,7 +19,6 @@ public class ChannelManager {
         this.defaultChannelType = defaultChannelType;
     }
 
-    // channelURI is type://id?key=value&key=value, or id with default channel
     public void notify(String channelURI, Alert alert, int alertCountSinceLastSent) {
         ChannelURI uri = parseChannelURI(channelURI);
         Channel channel = channels.get(uri.type == null ? defaultChannelType : uri.type);
@@ -27,22 +26,22 @@ public class ChannelManager {
         channel.notify(uri.id, uri.params, alert, alertCountSinceLastSent);
     }
 
+    // channelURI is type://id?key=value&key=value, or id with default channel
     ChannelURI parseChannelURI(String channelURI) {
-        ChannelURI uri = new ChannelURI();
-        final int ptcDelimiterLen = 3;
-        final int ptcIndex = channelURI.indexOf("://");
-        if (ptcIndex < 0) {
+        var uri = new ChannelURI();
+        final int protocolIndex = channelURI.indexOf("://");
+        if (protocolIndex < 0) {
             uri.id = channelURI;
             return uri;
         }
-        final int paramsIndex = channelURI.indexOf('?');
+        int paramsIndex = channelURI.indexOf('?');
         if (paramsIndex < 0) {
-            uri.type = channelURI.substring(0, ptcIndex);
-            uri.id = channelURI.substring(ptcIndex + ptcDelimiterLen);
+            uri.type = channelURI.substring(0, protocolIndex);
+            uri.id = channelURI.substring(protocolIndex + 3);
             return uri;
         }
-        uri.type = channelURI.substring(0, ptcIndex);
-        uri.id = channelURI.substring(ptcIndex + ptcDelimiterLen, paramsIndex);
+        uri.type = channelURI.substring(0, protocolIndex);
+        uri.id = channelURI.substring(protocolIndex + 3, paramsIndex);
         String parameters = channelURI.substring(paramsIndex + 1);
         uri.params = Arrays.stream(Strings.split(parameters, '&'))
             .map(entry -> Strings.split(entry, '='))