+++ ext/monitor/src/main/java/app/monitor/channel/ChannelURI.java
@@ -1,5 +1,6 @@
 package app.monitor.channel;
 
+import java.util.HashMap;
 import java.util.Map;
 
 /**
@@ -8,5 +9,5 @@ import java.util.Map;
 public class ChannelURI {
     public String type;
     public String id;
-    public Map<String, String> params;
+    public Map<String, String> params = new HashMap<>();
 }