+++ ext/monitor/src/main/java/app/monitor/channel/PagerDutyClient.java
@@ -1,6 +1,5 @@
 package app.monitor.channel;
 
-import app.monitor.PagerdutyConfig;
 import app.monitor.alert.Alert;
 import core.framework.api.http.HTTPStatus;
 import core.framework.http.ContentType;
@@ -18,57 +17,58 @@ import java.util.Map;
 /**
  * @author ajax
  */
-public class PagerdutyClient implements Channel {
+public class PagerDutyClient implements Channel {
     private static final String ACCIDENT_API_URL = "https://api.pagerduty.com/incidents";
-    private static final String FROM = "From";
     private static final String PRIORITY_ID = "priorityId";
     private static final String ESCALATION_POLICY_ID = "escalationPolicyId";
-    private final PagerdutyConfig pagerDutyConfig;
+    private final String token;
+    private final String from;
     private final HTTPClient httpClient;
 
-    public PagerdutyClient(HTTPClient httpClient, PagerdutyConfig pagerDutyConfig) {
+    public PagerDutyClient(HTTPClient httpClient, String token, String from) {
+        this.token = token;
+        this.from = from;
         this.httpClient = httpClient;
-        this.pagerDutyConfig = pagerDutyConfig;
     }
 
     @Override
     public void notify(String serviceId, Map<String, String> params, Alert alert, int alertCountSinceLastSent) {
-        PDCreateAccidentAPIRequest createRequest = createRequest(serviceId, params, alert);
-        HTTPRequest request = new HTTPRequest(HTTPMethod.POST, ACCIDENT_API_URL);
-        request.headers.put(HTTPHeaders.AUTHORIZATION, "Token token=" + pagerDutyConfig.token);
-        request.headers.put(FROM, pagerDutyConfig.from);
-        request.body(JSON.toJSON(createRequest), ContentType.APPLICATION_JSON);
+        PagerDutyAccidentAPIRequest pagerDutyRequest = createRequest(serviceId, params, alert);
+        var request = new HTTPRequest(HTTPMethod.POST, ACCIDENT_API_URL);
+        request.headers.put(HTTPHeaders.AUTHORIZATION, "Token token=" + token);
+        request.headers.put("From", from);
+        request.body(JSON.toJSON(pagerDutyRequest), ContentType.APPLICATION_JSON);
         HTTPResponse response = httpClient.execute(request);
         if (response.statusCode != HTTPStatus.CREATED.code) {
             throw new Error(Strings.format("failed to trigger accident to pagerduty, status={}, response={}", response.statusCode, response.text()));
         }
     }
 
-    private PDCreateAccidentAPIRequest createRequest(String serviceId, Map<String, String> params, Alert alert) {
-        PDCreateAccidentAPIRequest createRequest = new PDCreateAccidentAPIRequest();
-        createRequest.incident = new PDCreateAccidentAPIRequest.Incident();
-        createRequest.incident.type = "incident";
-        createRequest.incident.title = Strings.format("[{}] {}", ASCII.toUpperCase(alert.app), alert.errorMessage);
-        createRequest.incident.service = new PDCreateAccidentAPIRequest.Service();
-        createRequest.incident.service.id = serviceId;
-        createRequest.incident.service.type = "service_reference";
+    private PagerDutyAccidentAPIRequest createRequest(String serviceId, Map<String, String> params, Alert alert) {
+        PagerDutyAccidentAPIRequest request = new PagerDutyAccidentAPIRequest();
+        request.incident = new PagerDutyAccidentAPIRequest.Incident();
+        request.incident.type = "incident";
+        request.incident.title = Strings.format("[{}] {}", ASCII.toUpperCase(alert.app), alert.errorMessage);
+        request.incident.service = new PagerDutyAccidentAPIRequest.Service();
+        request.incident.service.id = serviceId;
+        request.incident.service.type = "service_reference";
         String priorityId = params.get(PRIORITY_ID);
         if (priorityId != null) {
-            createRequest.incident.priority = new PDCreateAccidentAPIRequest.Priority();
-            createRequest.incident.priority.id = priorityId;
-            createRequest.incident.priority.type = "priority_reference";
+            request.incident.priority = new PagerDutyAccidentAPIRequest.Priority();
+            request.incident.priority.id = priorityId;
+            request.incident.priority.type = "priority_reference";
         }
-        createRequest.incident.urgency = "high";
-        createRequest.incident.body = new PDCreateAccidentAPIRequest.Body();
-        createRequest.incident.body.type = "incident_body";
-        createRequest.incident.body.details = details(alert);
+        request.incident.urgency = "high";
+        request.incident.body = new PagerDutyAccidentAPIRequest.Body();
+        request.incident.body.type = "incident_body";
+        request.incident.body.details = details(alert);
         String escalationPolicyId = params.get(ESCALATION_POLICY_ID);
         if (escalationPolicyId != null) {
-            createRequest.incident.escalationPolicy = new PDCreateAccidentAPIRequest.EscalationPolicy();
-            createRequest.incident.escalationPolicy.id = escalationPolicyId;
-            createRequest.incident.escalationPolicy.type = "escalation_policy_reference";
+            request.incident.escalationPolicy = new PagerDutyAccidentAPIRequest.EscalationPolicy();
+            request.incident.escalationPolicy.id = escalationPolicyId;
+            request.incident.escalationPolicy.type = "escalation_policy_reference";
         }
-        return createRequest;
+        return request;
     }
 
     private String details(Alert alert) {