+++ ext/monitor/src/test/java/app/monitor/channel/ChannelManagerTest.java
@@ -1,7 +1,6 @@
 package app.monitor.channel;
 
 import app.monitor.alert.Alert;
-import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -10,6 +9,9 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.util.Map;
 
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.ArgumentMatchers.anyMap;
+import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.verify;
 
 /**
@@ -25,32 +27,33 @@ class ChannelManagerTest {
 
     @BeforeEach
     void createChannelManager() {
-        channelManager = new ChannelManager(Map.of("slack", slackChannel, "pager", pagerChannel), "slack");
+        channelManager = new ChannelManager(Map.of("slack", slackChannel, "pagerduty", pagerChannel), "slack");
     }
 
     @Test
     void parseChannelURI() {
-        ChannelURI uriWithType = channelManager.parseChannelURI("slack://channelId");
-        Assertions.assertEquals("slack", uriWithType.type);
-        Assertions.assertEquals("channelId", uriWithType.id);
+        ChannelURI uri = channelManager.parseChannelURI("slack://channelId");
+        assertThat(uri.type).isEqualTo("slack");
+        assertThat(uri.id).isEqualTo("channelId");
+        assertThat(uri.params).isNotNull();
 
-        ChannelURI uri = channelManager.parseChannelURI("channelId");
-        Assertions.assertNull(uri.type);
-        Assertions.assertEquals("channelId", uri.id);
+        uri = channelManager.parseChannelURI("channelId");
+        assertThat(uri.type).isNull();
+        assertThat(uri.id).isEqualTo("channelId");
+        assertThat(uri.params).isNotNull();
 
-        ChannelURI uriWithTypeAndParams = channelManager.parseChannelURI("pagerduty://serviceId?priorityId=mockPriorityId&escalationPolicyId=mockEscalationPolicyId");
-        Assertions.assertEquals("pagerduty", uriWithTypeAndParams.type);
-        Assertions.assertEquals("serviceId", uriWithTypeAndParams.id);
-        Assertions.assertEquals("mockPriorityId", uriWithTypeAndParams.params.get("priorityId"));
-        Assertions.assertEquals("mockEscalationPolicyId", uriWithTypeAndParams.params.get("escalationPolicyId"));
+        uri = channelManager.parseChannelURI("pagerduty://serviceId?priorityId=mockPriorityId&escalationPolicyId=mockEscalationPolicyId");
+        assertThat(uri.type).isEqualTo("pagerduty");
+        assertThat(uri.id).isEqualTo("serviceId");
+        assertThat(uri.params).containsEntry("priorityId", "mockPriorityId").containsEntry("escalationPolicyId", "mockEscalationPolicyId");
     }
 
     @Test
     void notifyWithPagerChannel() {
         var alert = new Alert();
-        channelManager.notify("pager://channelId", alert, 1);
+        channelManager.notify("pagerduty://serviceId", alert, 1);
 
-        verify(pagerChannel).notify("channelId", null, alert, 1);
+        verify(pagerChannel).notify(eq("serviceId"), anyMap(), eq(alert), eq(1));
     }
 
     @Test
@@ -58,6 +61,6 @@ class ChannelManagerTest {
         var alert = new Alert();
         channelManager.notify("channelId", alert, 1);
 
-        verify(slackChannel).notify("channelId", null, alert, 1);
+        verify(slackChannel).notify(eq("channelId"), anyMap(), eq(alert), eq(1));
     }
 }