+++ CHANGELOG.md
@@ -13,6 +13,7 @@
   > to make certain cache only uses in memory store, for rarely changed or sensitive data (must not reside in redis as clear text)
 * maven-repo: deleted all 6.x version except 6.13.9
   > recommend to upgrade to latest version
+* monitor: support pagerduty (thanks Ajax's contribution !!!)
 
 ### 7.9.3 (11/23/2021 - 12/10/2021)